* Refactor time formatting utilities into their own module * Some cleanup as a result of code inspection
